requirements.txt
requests
pandas
azure.storage.blob
duckdb